pyproject.toml
[project]
name = "flask-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "google-generativeai>=0.8.5",
]
